backend-complete/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic==0.25.1
snowflake-connector-python==3.4.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
PyJWT==2.8.1
passlib[bcrypt]==1.7.4
email-validator==2.1.0
python-multipart==0.0.6
shopify-python-api==13.0.0
gunicorn==21.2.0
